Move check_invalid_constraint_type to req_install.py
